test: move allowHTTP1 Expect: 100-continue tests to their own file
Relocate the four allowHTTP1 fallback Expect tests and their raw-socket
helpers out of node-http2.test.js into node-http2-allowhttp1-continue.test.ts,
matching the existing node-http2-*.test split. Keeps these fast tests isolated
from the long-running maxSessionMemory test in the main file.
